Internal: Size control break to pieces [ED-22389]
